microservices/redutor-imagem/requirements.txt
Flask==3.0.2
requests==2.32.3
boto3==1.34.125
Pillow==2.1.0